+++ lib/hash/fnv1A.ts
@@ -2,6 +2,9 @@ import {toString} from './utils';
 export const FNV_32 = 2166136261;
 export const FNV_64 = 1099511628211n;
 
+// 32-bit FNV prime (16777619)
+const FNV_PRIME = 16777619;
+
 /**
  * Convert a provided value into a Fowler-Noll-Vo 1A hash
  * For more info: https://tools.ietf.org/html/draft-eastlake-fnv-03
@@ -11,20 +14,17 @@ export const FNV_64 = 1099511628211n;
  */
 function fnv1A (data:unknown, offset:number = FNV_32):number {
     let hash = offset;
-    const normalized = toString(data);
+    const str = toString(data);
+    const len = str.length;
 
-    //  Calculate the hash of the normalized data by looping over each char
-    const len = normalized.length;
     for (let i = 0; i < len; i++) {
-        hash ^= normalized.charCodeAt(i);
+        hash ^= str.charCodeAt(i);
 
-        /**
-         * 32-bit FNV prime: 2**24 + 2**8 + 0x93 = 16777619
-         * Using bitshift for accuracy and performance. Numbers in JS suck
-         */
-        hash += (hash << 1) + (hash << 4) + (hash << 7) + (hash << 8) + (hash << 24);
+        // Math.imul Performs C-like 32-bit multiplication in one CPU instruction.
+        hash = Math.imul(hash, FNV_PRIME);
     }
 
+    // Force unsigned 32-bit integer
     return hash >>> 0;
 }